pyproject.toml
[project]
name = "egg-api"
version = "1.0.0"
description = "EGG-API: safe, normalized API facade for GLAM search backends"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "EGG-API Contributors" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "httpx>=0.27.0",
  "PyYAML>=6.0.1",
  "jinja2>=3.1.0",
  "prometheus-client>=0.20.0",
  "structlog>=24.1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "pytest-xdist>=3.6.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
  "pre-commit>=3.8.0",
  "types-PyYAML>=6.0.0",
  # Backport of the stdlib ``tomllib`` (stdlib since 3.11); pulled in
  # only for 3.10 so test_sprint9_release can read pyproject.toml.
  "tomli>=2.0.1; python_version < '3.11'"
]
otel = [
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-instrumentation-httpx>=0.48b0",
  "opentelemetry-exporter-otlp-proto-http>=1.27.0"
]
redis = [
  "redis>=5.0.0"
]

[project.scripts]
egg-api = "app.cli:main"

[tool.pytest.ini_options]
addopts = "-q --cov=app --cov-report=term-missing --cov-report=xml --cov-fail-under=80"
testpaths = ["tests"]
# Pin the repo root so ``tests._fakes`` (and any other test-shared module)
# resolves identically on every runner — the default rootdir heuristic
# is fragile under GitHub Actions / xdist.
pythonpath = ["."]

[tool.coverage.run]
source = ["app"]
branch = true
omit = ["app/__main__.py"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
app = ["admin_ui/static/*.css", "admin_ui/templates/*.html"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

# ---------------------------------------------------------------------------
# Lint + format + type-check
# ---------------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["build", "dist", ".venv"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "S",   # flake8-bandit (security)
  "SIM", # flake8-simplify
  "RUF"  # ruff-specific
]
ignore = [
  "E501",   # line too long (handled by black)
  "B008",   # function call in default argument (FastAPI pattern)
  "S101",   # use of assert (fine in tests)
  "S104",   # binding to 0.0.0.0 (allowed in container entrypoint)
  "SIM108"  # ternary rewrite (readability judgement)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S", "B", "SIM"]
"app/cli.py" = ["S603", "S607"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
combine-as-imports = true

[tool.mypy]
python_version = "3.10"
files = ["app"]
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = false
ignore_missing_imports = true
# Codebase is partially typed; tighten progressively.
check_untyped_defs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true
